+++ CHANGELOG.md
@@ -69,6 +69,7 @@ General Public License v2 (GPLv2). See LICENSES directory or go to
 ## Fixed
 - **Breaking**: "Remove backups older than" value was stored only as years ([#2460](https://github.com/bit-team/backintime/issues/2460))
 - Prevent crash in case a plugin fails ([#2447](https://github.com/bit-team/backintime/issues/2447))
+- Ensure restore window stay open while restore is running. (Dominic Maluski, @maluskid, [#2503](https://github.com/bit-team/backintime/issues/2503))
 - Include SSH_AUTH_SOCK in cron environment to enable SSH agent access (Dan Kortschak, [@kortschak](https://github.com/kortschak), [#2506](https://github.com/bit-team/backintime/issues/2506))
 - Schedule mode "Repeatedly" using "Hourly" units is now consistent with other units ([#2507](https://github.com/bit-team/backintime/issues/2507))
 - Crash when use Btrfs-subvolume as backup destination (Daidalos [@D4id4los](https://github.com/D4id4los), [#2487](https://github.com/bit-team/backintime/issues/2487))